add basic querying for blocks by name + type... uses file system and filtering, may be very slow
